Fix hide_score, hide_score_by_problem, and hide_work default values
On the ProblemSetDetail page when changing from a "homework" set to a
"test" set the defaults for hide_score, hide_score_by_problem, and
hide_work are not being set.  This results in warnings being displayed
and probably other bad results later.  This is because the defaults are
not set in the FIELD_PROPERTIES constant.  These were not set before,
and somehow it worked, but things have changed and now it doesn't.