Fix an incorrect route name on the stats page
The problem links on a set statistics page were not going to the correct
place because the route name was wrong.  The prolem links were going to
the general set stats page instead of the problem stats page because the
instructor_set_statistics route was accidentally used instead of the
instructor_problem_statics route.