Update the tag widget.
The javascript is updated with javascript that does not use jQuery.

Furthermore, the old widget that was directly displayed in the page and
was ugly as hell has been replaced with a button.  When the button is
clicked a dialog opens with a dialog which contains the tag editor.

If a problem has invalid tags, instead of those intrusive native alert
dialogs a toast is shown.  The user can continue to interact with the
page even while the toast is present.  The toast disappears after five
seconds on its own.