Make the PGProblemEditor show the problem submit buttons for source paths.
Currently if you edit an OPL or Contrib problem from the library browser
the submit buttons are not shown in the problem rendered in the page.

To fix this the javascript needs to check for both "problem" and
"source_path_for_problem_file" to determine when to show the submit
buttons.