Update the MAA logo with an svg image.
This svg was downloaded from
https://upload.wikimedia.org/wikipedia/en/8/8c/Mathematical_Association_of_America_logo.svg.
The fill color was changed to white and the stroke color added and set
to the original bluish fill color.  This was done to make this look good
with all of the current webwork2 themes.  This is essentially the same
logo that is currently shown on the MAA website, and is a more compact
logo.