Instead of retrieving the permissions for all users with one query to the database for each user, get them all in a single query.
From the develop branch to the feature/classlist-permissions branch the
timing for the load of the classlist editor page for a course with more
than 18,000 users increased from 240 milliseconds to 1.69 seconds.  That
is a considerable slow down.  The main reason for this is what is
changed here.  This brings the page load time back down to 549
milliseconds.  Of course those timings are from one page load, but
similar timings occur on repetition.

This is still twice as long as long of a load time as before.  Part of
this is because previously when loading a large class (more than 200
users) no permissions were loaded at first since only the permissions
for visible users were loaded.  Although, timing shows that with the
change suggested here the user and permission queries are about the same
as before.  Timing further shows that the real slow down is the sorting.
Previously only visible users were sorted.  Now all users are sorted
regardless of which are shown.