Fix a couple of encoding issues with the render_rpc endpoint.
First the output of render_to_string is a Mojo::ByteStream object.  As
such when it is then passed as the html key of the json object in
`render` of FormatRenderedProblem, it is not encoded.  So the
`to_string` method needs to be called on that to make sure it is encoded
properly for the json output.

Second, the raw output now needs to be utf8 encoded since Mojolicious
does not encode data output.