Remove the null byte hack.
In fact the WeBWorK::Form module is no longer needed at all.  Just use
the Mojolicious request params hash directly.

I thought that I was going to need to say that this is paired with
https://github.com/openwebwork/pg/pull/778 which removes null byte
usage from PG.  However, it turns out this works without that.  This
null byte handling just isn't needed.

One thing that I am not certain of is how to encode past answers that
come from arrays (for checkbox answers).  At this point this separates
entries by a space.  Previously a null byte was used.  The null byte
does not show up when viewing past answers.  Is there something better
than a space to use?  Note that separate answers for a problem are
separated by a tab character.