Pass the options isInstructor and forceScaffoldsOpen through with the render_rpc endpoint.
Since these options are not passed on through the form in the default
template, they are not preserved after submitting answers.  As such
problems with scaffolds come back after submitting with the latter
scaffolds grayed out and unable to be opened when a problem is rendered
on the set detail page, the problem sets editor page, or on the library
browser page.