Minor improvements and fixes for the attempts table.
Change the preview span to a preview div.  We make that a span, but set
it as display `block`.  If you want display block, generally you should
be using a div.  Furthermore, many answer previews now use divs for
layout which results in invalid html inside a span.

Also wrap the attempts table in a `table-responsive` div.  This cleans
up the page size on narrow screens or when the results table becomes
excessively large (as can quite frequently occur with messages and
correct answers).

Also fix display mode for MathJax equations in answer previews.  This
fixes issue #1921.