Suggestions and fixes.
When the problemGraderScore is set to Points, instead of adding the
percent input with label, help, and all of the html markup that will not
be used and should not be in the DOM, just add a hidden input for the
percent score in this case.

Also change some variable names to be more meaningful as to what they
represent.

Finally, fix alerts.  The bootstrap `alert` class adds padding and
margins and such that are not desired here and so the necessary css was
previously added in math4.scss.  However, bootstrap has changed the
`alert` class css and so the css added in math4.scss needs to be
updated.