Update the checkbox handling for the PG problem editor save form.
This makes it so that the 'Create backup' checkbox is checked when the
page loads initially, but maintains the status that the user set it to
on any page submission.  Currently it is simply always checked which is
not intuitive.

The 'Delete oldest backup' checkbox always maintains the current status
set by the user unless the maximum number of backups has been reached.
From then on it is always checked, and the user must manually uncheck
it.