Make the revert tab on the PG problem editor page always there.
However, it is disabled if there is nothing to revert to.

When a temporary file is saved by javascript, then it is enabled by
javascript.

If a backup file is made, the revert tab is also enabled.  However, it
only displays a message that the page needs to be reloaded to view the
backups that have been made.  If 'Take Action' is clicked the page
merely reloads.  That is unless a temporary file is being edited, in
which case the currently edited file is reverted to the original file.

Furthermore, the 'Delete oldest backup' check box is disabled when no
backups are initially detected, and is enabled when a backup is made.