Add DB support to allow inserting multiple UserSet records in one call and add basic transaction support and use it when assigning sets/problems to multiple users.
Add code to allow inserting multiple records for a single
user and set into the problem_user table as a batch, which
reduces the number of test calls made to the database,
and inserts them a batch so the prepared statement gets reused.

Make assignSetToGivenUsers so that assignSetsToUsers can use that
instead of looping over the users and calling assignSetToUser.

Add assignMultipleProblemsToGivenUsers which allows
inserting problems records for a given set to many users in one call.

Provide the setID and an array of problem_id values instead of an
array of globalProblem records.

Modify addUserMultipleProblems of lib/WeBWorK/DB.pm to use
   eval { $self->{problem_user}->insert_records([@problemList]); };
so that a single DB call makes all the inserts.
Test that no records for this user and set exist before making
that call.

These changes greatly speed up the process of "assign set to all users".

Have the UsersAssignedToSet code use the code to add users to
sets more efficiently (grouped DB operations).

Have the WebworkWebservice use the new methods when processing
assignSetToUsers and addProblem calls.