Update user information in cached list when saving a user in the Classlist Editor.
Since the database is not queried to retrieve updated information after
actions are performed the data for saved users needs to be updated in
the cached lists so that the updated information is displayed when the
templates are loaded later.

This fixes #1933.

Also fix sorting that was also broken in #1888.  The fallback methods
are needed in order for sorting to work correctly in case the first
three do not sort thing properly.  Currently, if you load the page with
users that do not have the information that the first three methods sort
on, then the users end up in a different order each time you load the
page.  For example, if using the default sort and users do not have
first or last names.