Rewrite third party asset handling.
This removes the third-party-assets.json file, and the contents of that
file are no longer added to the static-assets.json file.  Instead the
lib/WeBWorK/Utils.pm getAssetURL method loads the package.json file and
uses that to determine how to serve any file located in node_modules
(both from webwork2/htdocs and pg/htdocs).

The option thirdPartyAssetsUseCDN is added in site.conf.  If that is set
to 1, then getAssetURL constructs the jsdelivr CDN URL.  If that option
is set to 0, then the file is served from the node_modules directories,
and furthermore the version is appended as a URL parameter to the
node_modules path.