Move the page specific help link into the page title.
The question mark help icon in the main menu (the site navigation side
bar on the left) is not a good place.  It turns out many that have been
using WeBWorK for years don't even know there is page specific help that
this icon makes available.  So instead this icon is moved up by the page
title.  This will make it much more noticable and more clear as to what
it is offering help for.

Note that this is still not shown for students, and it is also not shown
if a page does not have a help file (the nohelp help file has been
removed).  In the latter case, that should be fixed and a help file
added.