Update instructor help files (team effort?)
Barebones help files have been added for pages that do not already have
help files.  These should be improved to actually say something useful.
Existing help files should be reviewed and improved as well.

When updating these help files remember two things.  These are not shown
to students, and these are only available when the page that the help is
for is already visible.  The second things means that these help files
should not just list the content of the page that is already visible
for the user.  That is pointless.  Instead the help files should
describe more detailed usage of the page, and most importantly, things
that are not intuitive from the layout of the page to begin with.