Add the PGML lab problem to the modelCourse and render locally for the PG problem editor.
The Rochester course that contains the PGML lab problem linked to for
the PG Problem Editor is not working anymore.  So the PGML lab link in
the problem editor needs to be handled differently.  This adds the
problem to the modelCourse so it will be copied into newly created
courses.  Then the PG problem editor checks to see if the file exists in
the course templates directory, and if so adds a link that opens the
problem in an undefined set and a new window.

This means that system administrators will need to update their local
modelCourse copy, and then newly created courses will get the file.  Old
coursed will need the file to be copied into their templates directory.
If the file PGMLLab/PGML-lab.pg is not present in the course templates
directory, the link will simply not be shown (better than a dead link to
a server that isn't operational).

Another option would be to add this problem to the OPL.  Then the
problem would be available for old courses assuming the OPL is updated.