Interpolate the `code` tags into the sentence in InstructorScoring.html.ep.
Also, remove the comment about what the "Success Index" in
InstructorStats.html.ep.  The success index has a different formulation
on the Stats page depending on if you are viewing the set as a whole, or
if you are viewing the stats for a single problem.  In addition there is
a popover directly in the page explaning the column, so there is no need
to mention it in the page help.