No more "Take Action!" button.  Stop taking action.  Just sleep.
On the pages that utilize action tabs, instead of the button reading
"Take Action!" it now has the text of the currently active tab.

Note that some of the help files reference the "Take Action!" button.
That will be changed in #1967.

Also, there are "Take Action!" buttons on the location management pages
in the course administration course.  Those are left for now.  Those
pages do not use the action tab javascript.