Fix invalid course logins.
Checking the param('user') isn't good enough to determine when to show
the help icon by the title.  When an invalid user attempts to sign in
that parameter will be set.  However, a user corresponding to that
parameter does not exist, and so the Authz hasPermissions call errors
out.  It is enough to check the `$authen->{was_verified}`.  If that is
true then there is a valid user as check_user has been called and found
a valid user from the value of param('user').