Don't allow renaming the current course (namely the admin course).
I have noticed that when in the admin course the admin course is listed
on the "Rename Course" page.  So I tried it out.  I renamed the admin
course, changing the courseID from admin to admin_blah.  The page
returned from the rename and said this course does not exist.  The
rename was of course successful, and the admin course no longer existed.
I then had to manually put everything back (rename the tables in the
database and rename the course directory).  This probably should not be
something that the page allows.  So this pull request removes the
current course (which would be the admin course) from the list.  This is
the same way that this is done for the "Delete Course" page.