Also don't list the admin course on the "Archive Course" page.
There is not point in listing the admin course on this page, and then
giving an error saying "You can not archive the course you are using."

Although, should we allow archiving of the admin course?  It works
perfectly fine to archive the course you are using.  In fact we even
have a link to do that in the site nav for a regular course.  The link
is visible anytime you are on the "File Manager" page.