Use the empty string for the "None" values and don't save "None" to the database (and other clean up).
I don't like saving the "None" value to the database.  Furthermore, it
is almost always the case that a default select value is better handled
with an empty value.  This is not an exception.

@somiaj:  Note that if you have 'None' saved in the database from when
you were saving this to the database, you will get a warning on the page
about that file not existing.  That will go away once you toggle
settings and it is removed from the database.

Clean up some of the mess of `if (defined $var && $var)`
conditionals.  For values that are expected to one of undef, the empty
string, or some nonempty string it is equivalent and much cleaner to
just check `if ($var)`.

Fix some of the 'Save as' messages.  Currently if no filename is
specified you get the message
	No filename was specified for saving! The message was not saved.
and then you also are given the message
	Invalid file name "FIXME no output file specified". All email
	file names must end with the ".msg" extension. Choose a file
	name with the ".msg" extension. The message was not saved.
which really doesn't make sense.

Remove the setting of $c->{message}.  It turns out that was still set.
The display of that message has been deleted, and it really isn't needed
anymore.

Also remove some unused variables.