Remove the allowedRecipients option.
This option no longer is used for anything now that PG problems can not
send email anymore.

Also clean up the generated course.conf file.  The segment regarding
$mail{feedbackRecipients} is also no longer added to the course.conf
file.  That can be set from the course configuration (written to
simple.conf), and is not needed here.  The actual setting is usually not
set here anyway.  All the commentary about this course environment
setting is certainly not needed here.  The copyright and license
certainly do not need to be output to a generated file.  Also remove
some excessive newlines and make the output more condensed.

The course.conf file in courses.dist/modelCourse is modified
accordingly.  In addition, the commentary about the
$achievementExcludeSet variable is removed.  Really this file should be
deleted.  The file itself is not used for anything. The only thing that
it has of value is the documentation on the $viewable_sections and
$viewable_recitations hashes.  Those hashes should be able to be set
from the course configuration, and the documenation added in
defaults.config.

For now I left the dbLayoutName and dbLayout setting in the course.conf.
Does anyone actually use the sql_moodle layout?  Is it even valid
anymore?  Should this be entirely removed from the code, and just have a
fixed database layout?