Make renaming a course or unarchiving a course to a new name work again.
This was using the old four argument form of the CourseEnvironment
constructor.  I missed that in #1985.

To test this rename a course or unarchive a course to a new name.  On
the current release candidate branch you will get an error.  With this
pull request it will work.

Note that if you test this on the release candidate branch and rename a
course, the course directory will be renamed.  So you will need to
manually rename that back to what it was.  The database tables will
still be under the old course name.  If you attempt to unarchive to a
new name, then the course directory will be created with the new name,
but the database tables will not be.  So you will need to manually
delete that new course directory.