Fix two column hardcopy themes to actually be two columns.
If the hardcopy set header happens to have the
BEGIN_ONE_COLUMN/END_ONE_COLUMN calls this is done by calling
`\twocolumn` with its header text starting and ending a
`@twocolumnfalse` section.  But if the set header does not call
BEGIN_ONE_COLUMN/END_ONE_COLUMN there is no `\twocolumn` call to start
two column mode.  To fix that the `\twocolumn` call needs to occur in
the hardcopyPreamble.tex file.

This fixes #1973.

Also it is time to remove that `\newpage` call at the beginning of the
document in all themes.  That is clearly added by someone that doesn't
know what they are doing with TeX.  There is no point in ending the
current page at the beginning of the document.

Additionally, there is no need to set the page number to 1 on the first
page of the document.  TeX does that anyway.