Make the achievement editor sublinks have the "instructor" color.
I noticed that if you are on the "Achievement Editor" page, and click on
the "Edit Users" column for an achievement, the achievment nav item that
is added with the achievement id does not have the "instcutor" coloring,
and so stands out oddly.  This adds the
`list-group-item list-group-item-primary` classes to make that match the
other instructor links.

Also fix some lines that have become overly long in that file.