Another tweak to the pg htdocs url configuration.
If `$pg_htdocs_url` is not set in `site.conf`, then pages load due to a
fallback in `lib/Mojolicious/WeBWorK.pm`.  However, if a problem uses
pg javascript the problem fails to load and there are lots of errors.
This adds a fallback to the course environment as well so that those
problems still load.