More shebang clean up.
Remove the shebang from perl module files in bin.  Remove the -w flag in
the check_database_charsets.pl script.  The usage of that flag is
discouraged as it shows warnings for all perl used by the script, even
those system libraries not under your control.  Furthermore, flags don't
work in that way for the `#!/usr/bin/env perl` shebang.