When deleting the user currently being acted as, stop acting.
If you delete the user that you are currently acting as, the
confirmation page fails to load and an error is shown.  This makes it so
that if the effectiveUser no longer exists, it automatically stops
acting as that user.  This is detected in
templates/ContentGenerator/Base/login_status.html.ep, and in that case
it simply changes the effectiveUser parameter back to the current user
id.