Update the docker build with missing modules.
When I added the LTI 1.3 authentication, I forgot to add the modules for
that.

Remove the libcgi-pm-perl package.  That is not needed anymore!

Remove the libcrypt-ssleay-perl package as that is not needed.

Add the dependencies of Mojo::SQLite as that speeds up the cpan
installation.

Don't install Mojolicious::Plugin::NotYAMLConfig as that is now part of
Mojolicious directly, and so is included in the libmojolicious-perl
pacakge which is already installed.