Update the masthead layout.
This removes the masthead from bootstrap's grid layout, and instead used
display flex.  The div containing the webwork logo and sidebar menu
button has a width of 20%, the same as the sidebar.  The institution
logo is allowed to take up any space left from the login status to the
right.

Since the sidebar and this div have the same width, the sidebar no
longer overlaps with the institution logo part of the masthead giving a
better appearance with the new coloring of the "MAIN MENU" navbar brand.

Note that on narrow screens there should be no difference from before.