Show a message if invalid characters are used when creating a set instead of dying.
A faculty member at my institution just tried to create a set with a
colon in the name.  This resulted in an error.  Since I have the
MIN_HTML_ERRORS option turned on, all she got was "An error occured
while processing your request" with the error record identifier and
time.  This should not happen.  Instead this checks for invalid
characters before trying to craete the set and shows a message in that
case.

This has probably been this way for a long time.  I tested on 2.17, and
got the same error.