Restructure the PG error messages in Problem.pm
You can not have an `h3` in a `p` tag.

Make the first header for these warnings an `h2` tag since there
aren't any tags between this and the page title `h1`.  Also remove the
`style="color:red;"` from this, as that does not satisfy accessibility
contrast requirements with the background color here.