Show non fatal warnings in the PG problem editor.
These warnings are shown when generating a hardcopy in the problem
editor, but not when directly rendering the problem.  These warnings are
shown when you view the problem in the set for instructors, and should
certainly be shown for problem authors.  This was an oversight that this
parameter was not added to begin with.