Update the temporary directory comments and location.
The comment is updated to state that the directory chosen must be
have the correct permissions.  Note the comment already states that the
directory must be created.

The comment about the system default for apache2's document root is
incorrect.  On most systems that is `/var/www/html`, not `/var/www`.  So
that comment is corrected, and the example settings adjusted.

This is to address what is left of issue #2047.